+++ src/FastyBird/Connector/NsPanel/tests/fixtures/API/request/set_device_state.json
